refactor: Update AuthViewModel to use StateFlow for user and user details
Updated AuthViewModel to use StateFlow for managing the user and user details.

Removed the `user` property and replaced it with a `user` StateFlow.
Added a `userDetails` StateFlow to hold the user's details.
Added a `hasUser()` method to check if a user is signed in.
Updated the `onSignInResult` method to update the `user
` and `userDetails` StateFlows.
Initialized the `userDetails` StateFlow in the `init` block.