feat: Add bookmark search functionality
Added a search bar to the Recents screen that allows users to search for bookmarks.
The search results are displayed in a list, and the user can tap on a bookmark to open it in their browser.
Added search functionality to the BookmarkRepository and BookmarkDao.
Updated the Recents screen to display search results.
Added a new ScreenState enum to represent the current state of the Recents screen.
Added a new SearchScreenState data class to represent the state of the search bar.
Added a new SearchContent composable to display search results.